fix: 조회 기준 도메인과 반환 도메인이 달라 발생한 오류 수정
