+++ src/main/java/com/yeoljeong/tripmate/infrastructer/persistence/repository/PlanParticipationRepositoryImpl.java
@@ -11,8 +11,6 @@ import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Slice;
 import org.springframework.stereotype.Repository;
 
 @Repository
@@ -78,11 +76,6 @@ public class PlanParticipationRepositoryImpl implements PlanParticipationReposit
   @Override
   public boolean existsOpenPlan(UUID userId) {
     return jpaRepository.existsOpenPlan(userId);
-  }
-
-  @Override
-  public Slice<Plan> findMyParticipatedPlans(UUID userId, Pageable pageable) {
-    return jpaRepository.findMyParticipatedPlans(userId, pageable);
   }
 
   @Override
